Adding another change to use size_t

+++ components/scream/src/physics/shoc/atmosphere_macrophysics.hpp
@@ -416,7 +416,7 @@ protected:
   void set_computed_group_impl (const FieldGroup& group);
 
   // Computes total number of bytes needed for local variables
-  int requested_buffer_size_in_bytes() const;
+  size_t requested_buffer_size_in_bytes() const;
 
   // Set local variables using memory provided by
   // the ATMBufferManager
